Avoid the dependency of related models on import
This huge refactor proposes a new data architecture for the project in
order to softwn the dependency of related objects on data import.

This also removes old code that will be brought back to this new
architecture soon. Old migrations and database tables can be deleted as
they will be recreated.